[#1404] Separate custom harness implementations for std and no_std
