+++ iceoryx2-bb/container/tests-nostd/src/main.rs
@@ -69,4 +69,4 @@ impl Default for GlobalHeapAllocator {
 #[global_allocator]
 static GLOBAL: GlobalHeapAllocator = GlobalHeapAllocator::new();
 
-iceoryx2_bb_testing::bootstrap!();
+iceoryx2_bb_testing::test_harness!();
