+++ iceoryx2-bb/derive-macros/tests-nostd/src/main.rs
@@ -59,4 +59,4 @@ impl Default for GlobalHeapAllocator {
 #[global_allocator]
 static GLOBAL: GlobalHeapAllocator = GlobalHeapAllocator::new();
 
-iceoryx2_bb_testing::bootstrap!();
+iceoryx2_bb_testing::test_harness!();
