+++ iceoryx2-bb/lock-free/tests-nostd/src/main.rs
@@ -64,4 +64,4 @@ impl Default for GlobalHeapAllocator {
 #[global_allocator]
 static GLOBAL: GlobalHeapAllocator = GlobalHeapAllocator::new();
 
-iceoryx2_bb_testing::bootstrap!();
+iceoryx2_bb_testing::custom_harness!();
